Add support for numbers in the password
- Added support for the --numbers feature. The user can now use the --numbers flag to make the program generate a password that includes
both lowercase letters and numbers. Also cleaned up some comments and the displayHelp function will now provide the user with information
on how to use the --numbers flag. Finally, added an if statement to cover the rare chance that even when using the --numbers flag, it generates
a password without numbers. This if statement will remove the final character of the password and replace it with a number.